Move LegalInformation to a specific component
